Implemented a function to generate a single response
